@@ -4,15 +4,16 @@
   "private": true,
   "dependencies": {
     "@material-ui/core": "^4.12.4",
-    "axios": "^0.21.1",
-    "chart.js": "^2.9.4",
-    "react": "^17.0.2",
-    "react-chartjs-2": "^2.11.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3"
+    "axios": "^0.27.2",
+    "chart.js": "^3.9.1",
+    "react": "^18.2.0",
+    "react-chartjs-2": "^4.3.1",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "5.0.1"
   },
   "scripts": {
-    "start": "SET NODE_OPTIONS=--openssl-legacy-provider && react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
